parthenope: Compiler delint

git-svn-id: https://svn.aros.org/svn/aros/trunk@45182 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ arch/ppc-sam440/boot/parthenope/src/sfs.c
@@ -133,7 +133,7 @@ static struct BNode *get_bnode(sfs_boot_dev_t * self, uint32_t key)
 	void *buffer;
 	struct BNode *bnode, *tmp;
 	uint32_t next;
-	uint16_t i;
+	int i;
 
 	bnode = NULL;
 	buffer = malloc(self->blocksize);
